There is a bug when creating a new delivery session appointment along with session feedback. The asynchronous events that are triggered on feedback sumission are unable to obtain the current appointment because it's not been persisted at the time of execution as it's an asychronous event.
We could persist the appointment before sending the event - thereby it would be able to extract the appointment; however we are sending the delivery session entity to the event handler.

There is a possibility that the transaction may have been closed when handling the event, so any lazy loading would fail with Hibernate exceptions.

A way around this is to transform any necessary entities into DTOs - this has a significant knock on effect, so all code changes in this commit are connected to this specific goal.