This commit aims to fix:
1. Issues arising from rebasing from main
2. Modifications to endpoint on back-end.

This involves changing ActionPlanAppointment | InitialAssessmentAppointment to just Appointment throughout because the DraftAppointment won't have an id but ActionPlanAppointment is expecting one now.
When extracting from the Draft we need to transform it into an ActionPlanAppointment but this is not possible if no id exists.

The changes before this commit assumes a single endpoint for scheduling/rescheduling an appointment. This has been split into sepearate flows for schedule and reschedule.

Reschedule expects a referralId, sessionNumber and appointmentId.
Schedule expects a referralId and sessionNumber.

As it currently stands this commit does not solve the above. Notably:

1, We still need to ensure that past appointments flow works for the new end points.
2. Some tests remain broken