Add form validation for appointment scheduling which validates the appointment time is not before *today*.
The reason for this is that the API does not currently support appointments before today, and if you
continue to submit the form you end up with an error page with a less useful error message. This
change brings the error to the users attention sooner so it can be corrected without the pain of
confirming and then being presented with an error page.