Only require feedback for appointments that happened on a previous day.
This replicates the current behaviour of the API, allowing the existing UI
to continue to function as normal without any changes.

Once the UI is allowing appointments to be created in the past, we can revert
this change to ensure that any past appointment, even if it was on the current
day, requires that attendance information be passed along side the scheduling
information.