user user id instead of username to look up service provider users in hmpps auth.
depending on how the username is used, email address changes in hmpps auth can cause
pages to become broken entirely in the UI (for example where looking up a referral
assignee with an invalid or old username).

this change removes the old hmpps auth service method for looking up records by username,
and replaces it with the equivalent call to look up users using user id. there are lots
of changes to the integration tests, but it's mainly just find and replace.