Sort pageable PP dashboard case lists using server side API params instead of client-side js.
This change strips out a tonne of javascript from the 'mojSortableTable' component
and allows the backend to apply the sort order across the full paged list of cases.
The selected sort order is stored in redis for a week (this length of time is
arbitrary and we will tune it later).