never pass 'undefined' to http-error createError function.
some errors thrown by 'RestClient' do not have a http status code,
for example timeouts or connection errors. in these cases,
pass a status of 500 instead. this doesn't really have any effect
on the application's behaviour, because 'external' errors always
end up with a 500 response anyway. this is a sensible default either
way because a 500 is the appropriate response code for an unhandled
internal error.

this also stops us getting alerts in sentry that look like
'argument #1 unsupported type undefined'.