Write referral details to the new update-able table alongside the existing referral table.
This change is step one in the roll out of referral details that are allowed to change
after a referral is sent. Currently the new table is write-only, but that means
that once we backfill the table using data from existing referrals, it will be then fully
up-to-date with no need to backfill anything again.