Switch to passive WAF monitoring
We are getting many complaints about genuine requests being blocked by
the web application firewall (WAF).

I cannot see the triggering rules in the logs for the UI, but I can see
them in the API.

For now, use the same config as the API so we can monitor rules
triggered and create a reasonable allow-list.