Recover from errors on session/draft storage
So that intermittent redis connection errors will not get the
application in a "stuck" state.

The `redisClient` in `hmppsAuthService.ts` recovers from all
intermittent outages but this main one does not and the only difference
is the `on('error', ...)` handler.

Tested locally and without this handler, the connection remains dead.
With it, it recovers.

(This does not replace a proper liveness check though -- that one should
ping all redis connections)