Invalidate client token cache on client rotation
We are having an issue with 401 Unauthorized responses when UI makes
hmpps-auth calls.

We think it's due to this token being cached in Redis, after a token
rotation happened this morning.

To mitigate, we should cache by the client ID in they cache key, so any
further rotation will discard the previous cache value.